requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.2
albumentations>=1.3.0
opencv-python
scikit-learn
matplotlib
seaborn
tqdm
numpy
pandas
transformers>=4.30.0
pillow>=9.0.0
streamlit>=1.28.0
plotly>=5.18.0
